requirements.txt
async-timeout==3.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
msgpack==0.6.2
msgpack-python==0.5.6
pymetasploit3==1.0.2
python-nmap==0.6.1
requests==2.22.0
urllib3==1.25.7
